// src/app/core/models/file.ts
export interface Upload {
    progress: number;
    state: 'PENDING' | 'IN_PROGRESS' | 'DONE';
    content?: any;
}